demo/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.12",
    "django-book-manager>=0.3.2",
    "django-cognito-m2m[drf]",
    "djangorestframework>=3.17.1",
]

[tool.uv.sources]
django-cognito-m2m = {path = "..", editable = true}